% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_annotation.R
\name{add_gene_names_to_granges}
\alias{add_gene_names_to_granges}
\title{Add names of genes local to ranges stored in GenomicRanges object}
\usage{
add_gene_names_to_granges(granges, txdb, flank_length = NULL)
}
\arguments{
\item{granges}{GenomicRanges object}

\item{txdb}{TxDb containing annotations}

\item{flank_length}{width of flanking region to add to each range}
}
\value{
GenomicRanges with gene names added
}
\description{
Add names of genes local to ranges stored in GenomicRanges object
}
