% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/back_to_back_manhattan.R
\name{back_to_back_manhattan}
\alias{back_to_back_manhattan}
\title{Writes out a back-to-back Manhattan plot.}
\usage{
back_to_back_manhattan(
  top_gRanges,
  bottom_gRanges,
  top_label,
  bottom_label,
  main,
  ymax = 15,
  tick_dist = 1e+05,
  axis_label_margin = 0.03,
  main_title_cex = 2.7,
  axis_label_cex = 1.8,
  axis_label_offset = 0,
  axis_tick_cex = 1,
  plot_genes = F,
  gene_names_cex = 1,
  chrom_names_cex = 2,
  third_gRanges = NULL,
  third_label = NULL,
  chromosomes = NULL,
  zoom = NULL,
  plot_params = getDefaultPlotParams(plot.type = 4),
  top_gRanges_points.col = "2blues",
  bottom_gRanges_points.col = "2blues",
  third_gRanges_points.col = "2blues",
  points.cex = 1,
  top_highlight = NULL,
  bottom_highlight = NULL,
  third_highlight = NULL,
  top_highlight.col = "greenyellow",
  bottom_highlight.col = "greenyellow",
  third_highlight.col = "greenyellow"
)
}
\arguments{
\item{top_gRanges}{GenomicRanges object containing data for top plot}

\item{bottom_gRanges}{GenomicRanges object containing data for bottom plot}

\item{top_label}{Axis label for top Manhattan plot of back-to-back pair}

\item{bottom_label}{Axis label for bottom Manhattan plot of back-to-back pair}

\item{main}{Main title}

\item{ymax}{Maximum value for y axis}

\item{tick_dist}{Distance in basepairs between chromosome diagram ticks}

\item{axis_label_margin}{Size of margin between axis and its label (distance between axis and axis label)}

\item{main_title_cex}{Scaling factor for main title}

\item{axis_label_cex}{Scaling factor for axis label}

\item{axis_label_offset}{Offset for axis label}

\item{axis_tick_cex}{Scaling factor for axis ticks}

\item{plot_genes}{Flag to have karyoploteR plot gene structures beneath Manhattan}

\item{gene_names_cex}{Scaling factor for gene labels}

\item{chrom_names_cex}{Scaling factor for chromosome labels}

\item{third_gRanges}{GenomicRanges object containing data for third plot to be placed above back-to-back Manhattans}

\item{third_label}{Axis label for third plot}

\item{chromosomes}{List of chromosomes to plot. Expects strings of the form 'chrx' where 'x' is from the set {1, 2, ..., 22, X, Y}}

\item{zoom}{Coordinates for interval to magnify. Expects a string of the form 'chrx:a-b' where x is the chromosome number a and b are basepair coordinates.}

\item{plot_params}{List of named plot parameters to pass to \code{plotKaryotype}}

\item{top_gRanges_points.col}{Colours used to plot the points in top_gRanges}

\item{bottom_gRanges_points.col}{Colours used to plot the points in bottom_gRanges}

\item{third_gRanges_points.col}{Colours used to plot the points in third_gRanges}

\item{points.cex}{Size of the point symbols}

\item{top_highlight}{(GRanges, character vector, logical vector or numeric vector) The points to highlight in a different color in the top plot. If a GRanges (or anythng accepted by toGRanges) the points overlapping these regions will be highlighted. Otherwise the points will be selected with data\link{highlight}. If NULL no point will be highlighted. (defaults to NULL)}

\item{bottom_highlight}{(GRanges, character vector, logical vector or numeric vector) The points to highlight in a different color in the bottom plot. If a GRanges (or anythng accepted by toGRanges) the points overlapping these regions will be highlighted. Otherwise the points will be selected with data\link{highlight}. If NULL no point will be highlighted. (defaults to NULL)}

\item{third_highlight}{(GRanges, character vector, logical vector or numeric vector) The points to highlight in a different color in the third plot. If a GRanges (or anythng accepted by toGRanges) the points overlapping these regions will be highlighted. Otherwise the points will be selected with data\link{highlight}. If NULL no point will be highlighted. (defaults to NULL)}

\item{top_highlight.col}{Top plot highlight colour}

\item{bottom_highlight.col}{Bottom plot highlight colour}

\item{third_highlight.col}{Third plot highlight colour}
}
\value{
KaryoPlot object
}
\description{
Main back-to-back plot can be supplemented with a third plot and with gene structure diagrams.
}
\details{
At present we require that the p-value column in a GenomicRanges column be labelled 'P' or 'p'.
}
