% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backToBackManhattan.R
\name{backToBackManhattan}
\alias{backToBackManhattan}
\title{Writes out a back-to-back Manhattan plot.}
\usage{
backToBackManhattan(
  topGRanges,
  bottomGRanges,
  topLabel,
  bottomLabel,
  main,
  outputFile = NULL,
  width = 1800,
  height = 1000,
  ymax = 15,
  tickDist = 1e+05,
  plotGenes = F,
  thirdGRanges = NULL,
  thirdLabel = NULL,
  chromosomes = NULL,
  zoom = NULL
)
}
\arguments{
\item{topGRanges}{GenomicRanges object containing data for top plot}

\item{bottomGRanges}{GenomicRanges object containing data for bottom plot}

\item{topLabel}{Axis label for top Manhattan plot of back-to-back pair}

\item{bottomLabel}{Axis label for bottom Manhattan plot of back-to-back pair}

\item{main}{Main title}

\item{outputFile}{Path to output file. If omitted, plots to the screen device.}

\item{width}{Width of plot in pixels for writing to file device}

\item{height}{Height of plot in pixels for writing to file device}

\item{ymax}{Maximum value for y axis}

\item{tickDist}{Distance in basepairs between chromosome diagram ticks}

\item{plotGenes}{Flag to have karyoploteR plot gene structures beneath Manhattan}

\item{thirdGRanges}{GenomicRanges object containing data for third plot to be placed above back-to-back Manhattans}

\item{thirdLabel}{Axis label for third plot}

\item{chromosomes}{List of chromosomes to plot. Expects strings of the form 'chrx'.}

\item{zoom}{Coordinates for interval to magnify. Expects a string of the form 'chrx:a-b' where x is the chromosome number a and b are basepair coordinates.}
}
\description{
Main back-to-back plot can be supplemented with a third plot and with gene structure diagrams.
}
\details{
Requires at least two GenomicRanges objects for the main back-to-back plot. A third can be added on its own as a reference.
}
