% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bivariate_ecdf_cpp}
\alias{bivariate_ecdf_cpp}
\title{Bivariate empirical cdf}
\usage{
bivariate_ecdf_cpp(u_ref, v_ref)
}
\arguments{
\item{u_ref}{NumericVector reference points in first dimension}

\item{v_ref}{NumericVector reference points in second dimension}
}
\value{
NumericVector of estimated quantiles of specified pairs
}
\description{
This function fits a bivariate ecdf to the specified pair of variables and then evaluates the ecdf at the same points.
}
\details{
Assumes u_ref and v_ref specify pairs of points in the correct order.
}
\author{
Tom Willis
}
