% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bivariate_ecdf_lw_cpp}
\alias{bivariate_ecdf_lw_cpp}
\title{Bivariate empirical cdf using novel divide-and-conquer algorithm}
\usage{
bivariate_ecdf_lw_cpp(u_ref, v_ref)
}
\arguments{
\item{u_ref}{NumericVector reference points in first dimension}

\item{v_ref}{NumericVector reference points in second dimension}
}
\value{
NumericVector of estimated quantiles of specified pairs
}
\description{
Bivariate empirical cdf using novel divide-and-conquer algorithm
}
\author{
Tom Willis
}
