% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_maf_table.R
\name{compute_maf_table}
\alias{compute_maf_table}
\title{Returns table of MAF values by interval.}
\usage{
compute_maf_table(maf, breaks = seq(0, 0.5, length = 51), proportions = F)
}
\arguments{
\item{maf}{Vector of MAF values}

\item{breaks}{Vector of breaks to define intervals for table}

\item{proportions}{Flag to return table with proportions rather than counts}
}
\value{
Table
}
\description{
Returns table of MAF values by interval.
}
