% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_annotation.R
\name{get_exon_id_for_granges}
\alias{get_exon_id_for_granges}
\title{Fetch exon ID from TxDb for GenomicRanges object}
\usage{
get_exon_id_for_granges(granges, txdb)
}
\arguments{
\item{granges}{GenomicRanges object}

\item{txdb}{TxDb containing annotations}
}
\description{
Fetch exon ID from TxDb for GenomicRanges object
}
