% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_annotation.R
\name{get_gene_names_for_granges}
\alias{get_gene_names_for_granges}
\title{Fetch genes from TxDb for GenomicRanges object}
\usage{
get_gene_names_for_granges(granges, txdb, flank_length = NULL)
}
\arguments{
\item{granges}{GenomicRanges object}

\item{txdb}{TxDb containing annotations}

\item{flank_length}{width of flanking region to add to each range}
}
\description{
Fetch genes from TxDb for GenomicRanges object
}
\details{
Adapted from http://www.bioinsteps.com/2019/08/annotate-genomic-coordinates-using-r.html
}
