% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{gps}
\alias{gps}
\title{Genome-wide pairwise association signal sharing test statistic}
\usage{
gps(u, v)
}
\arguments{
\item{u}{Vector of p-values}

\item{v}{Vector of p-values}
}
\value{
Numeric value of statistic
}
\description{
Implementation of GPS test statistic from Li et al. (2015), 'Meta-analysis of shared genetic architecture across ten pediatric autoimmune diseases'
}
