% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{gps_maximands}
\alias{gps_maximands}
\title{Vector of GPS statistic maximands}
\usage{
gps_maximands(u, v)
}
\arguments{
\item{u}{Vector of p-values}

\item{v}{Vector of p-values}
}
\value{
Vector of maximands
}
\description{
Vector of GPS statistic maximands
}
