% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_test_stat.R
\name{gps_test_stat}
\alias{gps_test_stat}
\title{Genome-wide pairwise association signal statistic}
\usage{
gps_test_stat(u, v)
}
\arguments{
\item{u}{Vector of p-values}

\item{v}{Vector of p-values}
}
\value{
Numeric value of statistic
}
\description{
Genome-wide pairwise association signal statistic
}
