% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_to_granges.R
\name{gwas_to_granges}
\alias{gwas_to_granges}
\title{Convert GWAS data in data.table to GenomicRanges object}
\usage{
gwas_to_granges(data_table, bp_label = "BP38", chr_label = "CHR38")
}
\arguments{
\item{data_table}{data.table containing GWAS data}

\item{bp_label}{name of basepair column}

\item{chr_label}{name of chromosome column}
}
\value{
GenomicRanges object
}
\description{
Convert GWAS data in data.table to GenomicRanges object
}
