% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rds_or_run.R
\name{load_rds_or_run}
\alias{load_rds_or_run}
\title{Check for RDS file and if absent, make it}
\usage{
load_rds_or_run(file_path, function_to_run)
}
\arguments{
\item{file_path}{Path to file}

\item{function_to_run}{Function to generate object}
}
\value{
value of function_to_run
}
\description{
Check for RDS file and if absent, make it
}
