% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{mix_rexp}
\alias{mix_rexp}
\title{Generates samples from two-part mixture of exponentials}
\usage{
mix_rexp(n, rates = c(5, 1), alt_weight = 0.01, pval_scale = F)
}
\arguments{
\item{n}{Sample size}

\item{rates}{Rate parameter for each exponential component}

\item{alt_weight}{Mixing weight for first exponential component; second has weight 1-alt_weight}

\item{pval_scale}{Flag to return sample on \link{0,1} scale}
}
\value{
Sample from the mixture distribution
}
\description{
Generates samples from two-part mixture of exponentials
}
