% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrack_manhattan.R
\name{multitrack_manhattan}
\alias{multitrack_manhattan}
\title{Writes out a multitrack Manhattan.}
\usage{
multitrack_manhattan(
  gRanges,
  axis_labels,
  main,
  ymax = 15,
  axis_label_margin = 0.03,
  main_title_cex = 2.7,
  axis_label_cex = 1.8,
  axis_label_offset = 0,
  axis_tick_cex = 1,
  chrom_names_cex = 2,
  chromosomes = NULL,
  chrom_tick_dist = 1e+06,
  chrom_tick_cex = 1,
  plot_params = getDefaultPlotParams(plot.type = 4),
  track_margin = 0.06,
  points.col = "2blues",
  points.cex = 1,
  highlights = NULL,
  highlight.cols = NULL
)
}
\arguments{
\item{gRanges}{List of \code{GenomicRanges} objects to plot}

\item{axis_labels}{List of axis labels}

\item{main}{Main title}

\item{ymax}{Maximum value for y axes}

\item{axis_label_margin}{Size of margin between axis and its label (distance between axis and axis label)}

\item{main_title_cex}{Scaling factor for main title}

\item{axis_label_cex}{Scaling factor for axis label}

\item{axis_label_offset}{Offset for axis label}

\item{axis_tick_cex}{Scaling factor for axis ticks}

\item{chrom_names_cex}{Scaling factor for chromosome labels}

\item{chromosomes}{List of chromosomes to plot. Expects strings of the form 'chrx' where 'x' is from the set {1, 2, ..., 22, X, Y}}

\item{chrom_tick_dist}{Distance in bp between ideogram ticks}

\item{chrom_tick_cex}{Scaling factor for ideogram ticks}

\item{plot_params}{List of named plot parameters to pass to \code{plotKaryotype}}

\item{track_margin}{Proportion of track assigned to track margin}

\item{points.col}{Colours used to plot points}

\item{points.cex}{Size of the point symbols}

\item{highlights}{List of GRanges, or character, numeric, or logical vectors giving the points to highlight in a different color in the top plot. If a GRanges (or anythng accepted by toGRanges) the points overlapping these regions will be highlighted. Otherwise the points will be selected with data\link{highlight}. If NULL no point will be highlighted. (defaults to NULL)}

\item{highlight.cols}{Colours for each highlight}
}
\value{
KaryoPlot
}
\description{
Writes out a list of \code{GenomicRanges} objects in a stacked Manhattan plot.
}
