% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.R
\name{rgps}
\alias{rgps}
\title{Generates samples from the null distribution of the GPS statistic using samples from the exponential mixture}
\usage{
rgps(n, no_snps, rates = c(5, 1), alt_weight = 0.01)
}
\arguments{
\item{n}{Sample size}

\item{no_snps}{Number of SNPs/p-values to generate for each simulated GWAS data set}

\item{rates}{Rate parameter for each exponential component}

\item{alt_weight}{Mixing weight for first exponential component; second has weight 1-alt_weight}
}
\value{
Sample from the GPS null distribution
}
\description{
Generates samples from the null distribution of the GPS statistic using samples from the exponential mixture
}
