% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianMixture.R
\name{rmixture}
\alias{rmixture}
\title{Sample from two-Gaussian mixture.}
\usage{
rmixture(n, pi, mean, sd)
}
\arguments{
\item{n}{Size of sample}

\item{pi}{Mixing weight for standard normal component}

\item{mean}{Mean for non-standard normal component}

\item{sd}{Standard deviation for non-standard normal component}
}
\value{
Sample from mixture process
}
\description{
Draws a sample from a mixture of a standard normal and a non-standard normal with specified mean and standard deviation.
}
\details{
Mixing weights are specified by first element of parameter vector, say pi, as pi and 1-pi.
}
\examples{

sam<-rmixture(1e3, 0.1, 2, 3)

}
