% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cl_cfdr.R
\name{run_cl_cfdr}
\alias{run_cl_cfdr}
\title{Run cFDR program with command-line arguments}
\usage{
run_cl_cfdr(cl_args, PID_ROOT = Sys.getenv("pidRoot"))
}
\arguments{
\item{cl_args}{Vector of command-line arguments}

\item{PID_ROOT}{Path to root of PID project directory tree}
}
\description{
Run cFDR program with command-line arguments
}
