% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cl_fcfdr.R
\name{run_cl_fcfdr}
\alias{run_cl_fcfdr}
\title{Run fcFDR program with command-line arguments}
\usage{
run_cl_fcfdr(cl_args, PID_ROOT = Sys.getenv("pidRoot"))
}
\arguments{
\item{cl_args}{Vector of command-line arguments}

\item{PID_ROOT}{Path to root of PID project directory tree}
}
\description{
Run fcFDR program with command-line arguments
}
