% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cl_fcfdr.R
\name{run_cl_fcfdr_with_log}
\alias{run_cl_fcfdr_with_log}
\title{Run fcFDR program with command-line arguments and logging of independent SNP indices and indices of subset.}
\usage{
run_cl_fcfdr_with_log(cl_args, PID_ROOT = Sys.getenv("pidRoot"))
}
\arguments{
\item{cl_args}{Vector of command-line arguments}

\item{PID_ROOT}{Path to root of PID project directory tree}
}
\description{
Run fcFDR program with command-line arguments and logging of independent SNP indices and indices of subset.
}
