% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_qqplot.R
\name{stratified_qqplot}
\alias{stratified_qqplot}
\title{Writes out a stratified Q-Q plot.}
\usage{
stratified_qqplot(
  dataFrame,
  principalValueLabel,
  conditionalValueLabel = NULL,
  thresholds = c(1, 0.1, 0.01, 0.001, 1e-04)
)
}
\arguments{
\item{dataFrame}{\code{data.frame} containing p-values and conditional p-values}

\item{principalValueLabel}{label of principal p-value column in \code{dataFrame}}

\item{conditionalValueLabel}{label of conditional trait column in \code{dataFrame}}

\item{thresholds}{threshold values to define strata}
}
\value{
ggplot object
}
\description{

}
\details{
Note that this function does not do the heavy lifting of styling the plot's aesthetics.
}
