% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backToBackManhattan.R
\name{backToBackManhattan}
\alias{backToBackManhattan}
\title{Writes out a back-to-back Manhattan plot.}
\usage{
backToBackManhattan(
  topGRanges,
  bottomGRanges,
  outputFile,
  topLabel,
  bottomLabel,
  main,
  width = 1800,
  height = 1000,
  ymax = 15,
  tickDist = 1e+05,
  plotGenes = F,
  thirdGRanges = NULL,
  thirdLabel = NULL,
  chromosomes = NULL,
  zoom = NULL
)
}
\arguments{
\item{topGRanges}{GenomicRanges object containing data for top plot}

\item{bottomGRanges}{GenomicRanges object containing data for bottom plot}

\item{plotGenes}{}

\item{zoom}{}
}
\description{

}
\details{

}
