% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backToBackManhattan.R
\name{backToBackManhattan}
\alias{backToBackManhattan}
\title{Writes out a back-to-back Manhattan plot.}
\usage{
backToBackManhattan(
  topGRanges,
  bottomGRanges,
  topLabel,
  bottomLabel,
  main,
  ymax = 15,
  tickDist = 1e+05,
  axisLabelMargin = 0.03,
  mainTitleCex = 2.7,
  axisLabelCex = 1.8,
  plotGenes = F,
  geneNamesCex = 1,
  chromNamesCex = 2,
  thirdGRanges = NULL,
  thirdLabel = NULL,
  chromosomes = NULL,
  zoom = NULL,
  plotParams = getDefaultPlotParams(plot.type = 4)
)
}
\arguments{
\item{topGRanges}{GenomicRanges object containing data for top plot}

\item{bottomGRanges}{GenomicRanges object containing data for bottom plot}

\item{topLabel}{Axis label for top Manhattan plot of back-to-back pair}

\item{bottomLabel}{Axis label for bottom Manhattan plot of back-to-back pair}

\item{main}{Main title}

\item{ymax}{Maximum value for y axis}

\item{tickDist}{Distance in basepairs between chromosome diagram ticks}

\item{axisLabelMargin}{Size of margin between axis and its label}

\item{mainTitleCex}{Scaling factor for main title}

\item{axisLabelCex}{Scaling factor for axis label}

\item{plotGenes}{Flag to have karyoploteR plot gene structures beneath Manhattan}

\item{geneNamesCex}{Scaling factor for gene labels}

\item{chromNamesCex}{Scaling factor for chromosome labels}

\item{thirdGRanges}{GenomicRanges object containing data for third plot to be placed above back-to-back Manhattans}

\item{thirdLabel}{Axis label for third plot}

\item{chromosomes}{List of chromosomes to plot. Expects strings of the form 'chrx' where 'x' is from the set {1, 2, ..., 22, X, Y}}

\item{zoom}{Coordinates for interval to magnify. Expects a string of the form 'chrx:a-b' where x is the chromosome number a and b are basepair coordinates.}

\item{plotParams}{List of named plot parameters to pass to \code{plotKaryotype}}
}
\description{
Main back-to-back plot can be supplemented with a third plot and with gene structure diagrams.
}
\details{
At present we require that the p-value column in a GenomicRanges column be labelled 'P' or 'p'.
}
