% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multitrack_manhattan.R
\name{multitrack_manhattan}
\alias{multitrack_manhattan}
\title{Writes out a multitrack Manhattan.}
\usage{
multitrack_manhattan(
  gRanges,
  axis_labels,
  main,
  ymax = 15,
  axis_label_margin = 0.03,
  main_title_cex = 2.7,
  axis_label_cex = 1.8,
  axis_label_offset = 0,
  axis_tick_cex = 1,
  chrom_names_cex = 2,
  chromosomes = NULL,
  chrom_tick_dist = 1e+06,
  chrom_tick_cex = 1,
  plot_params = getDefaultPlotParams(plot.type = 4),
  track_margin = 0.06
)
}
\arguments{
\item{gRanges}{List of \code{GenomicRanges} objects to plot}

\item{axis_labels}{List of axis labels}

\item{main}{Main title}

\item{ymax}{Maximum value for y axes}

\item{axis_label_margin}{Size of margin between axis and its label}

\item{main_title_cex}{Scaling factor for main title}

\item{axis_label_cex}{Scaling factor for axis label}

\item{axis_label_offset}{Offset for axis label}

\item{axis_tick_cex}{Scaling factor for axis ticks}

\item{chrom_names_cex}{Scaling factor for chromosome labels}

\item{chromosomes}{List of chromosomes to plot. Expects strings of the form 'chrx' where 'x' is from the set {1, 2, ..., 22, X, Y}}

\item{chrom_tick_dist}{Distance in bp between ideogram ticks}

\item{chrom_tick_cex}{Scaling factor for ideogram ticks}

\item{plot_params}{List of named plot parameters to pass to \code{plotKaryotype}}

\item{track_margin}{Proportion of track assigned to track margin}
}
\value{
KaryoPlot
}
\description{
Writes out a list of \code{GenomicRanges} objects in a stacked Manhattan plot.
}
